Refactor SceneController into modular sub-controllers
