+++ ui/scene/__init__.py
@@ -1,1 +1,14 @@
-"""This package contains modules related to scene management."""
+"""Modules related to scene management."""
+
+from .scene_controller import SceneController
+from .puppet_ops import PuppetOps
+from .object_ops import ObjectOps
+from .library_ops import LibraryOps, LibraryPayload
+
+__all__ = [
+    "SceneController",
+    "PuppetOps",
+    "ObjectOps",
+    "LibraryOps",
+    "LibraryPayload",
+]
